refactor: rename api-client.ts to broker-client.ts
